gcp/requirements.txt
flask
torch
transformers
fuzzywuzzy
google-cloud-storage
pandas

# torch==1.9.0
# transformers==4.11.3
# fuzzywuzzy==0.18.0
# pandas==1.3.4
# google-cloud-storage==1.41.0
